New feature: fetching and storing of raw HTML from http://www.vybory.izbirkom.ru/
New rake task has been added: grab:fetch_raw_html All the data is fetched
from http://www.vybory.izbirkom.ru/ and stored in the file system.

TODO: add rake task to process these files and fill the database.

Reasoning: Fetching of the data from http://www.vybory.izbirkom.ru/ is a
very slow and sometimes painfule process (depends on internet
connectivity). The need to fill the data from the network slows down the
development. It also can be useful in production environment as the
election's committee's web site will be overloaded with requests and the
data will be updated there. Better to have prepared HTML files which are
ready-to-use rather than fetch the data directly from the network to
the database. It also can be used for archiving purposes (in case
election committee will change the data on the same page)

Russian translation is below:

Добавлен новый task для rake'a: grab:fetch_raw_html Производится опрос
сайтов избирательных комиссий и все данные, необходимые для БД,
сохраняются на файловой системе.

TODO: Добавить task для обработки файлов с html-данными и загрузки их в БД.

Обоснование: Процедура опроса сайтов избиркомов и скачивания данных
может занимать значительное время. Это замедляет процесс разворачивания
dev. окружения. Эти данные также могут быть полезны в  целях
архивирования исходных данных избиркомов.